feat : [step2] 주문하기 받는사람섹션 기본 UI 설계
