--- src/components/ProtectedRoute.tsx
@@ -1,20 +1,0 @@
-import { type ReactNode } from 'react';
-import { Navigate, useLocation } from 'react-router-dom';
-import { useAuth } from '@/contexts/AuthContext';
-
-interface ProtectedRouteProps {
-  children: ReactNode;
-}
-
-const ProtectedRoute = ({ children }: ProtectedRouteProps) => {
-  const { isLoggedIn } = useAuth();
-  const location = useLocation();
-
-  if (!isLoggedIn) {
-    return <Navigate to="/login" state={{ from: location.pathname }} replace />;
-  }
-
-  return <>{children}</>;
-};
-
-export default ProtectedRoute;
