+++ src/components/organisms/ReceiverSection/styles.ts
@@ -9,37 +9,39 @@ export const Container = styled.div`
   border: 1px solid ${({ theme }) => theme.semantic.background.fill};
 `;
 
+export const FirstSection = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding-bottom: ${({ theme }) => theme.spacing.spacing3};
+`;
+
 export const SectionTitle = styled.div`
   display: flex;
-  justify-content: flex-start;
-  padding-bottom: ${({ theme }) => theme.spacing.spacing3};
+  justify-content: space-between;
+  align-items: center;
   ${({ theme }) => theme.typography.subtitle1Bold};
 `;
 
-export const FormContent = styled.div`
-  width: 100%;
+export const ModalBtn = styled.button`
+  box-sizing: border-box;
+  ${({ theme }) => theme.typography.label1Regular};
+  border : 1px solid ${({ theme }) => theme.semantic.border.default};
+  border-radius: ${({ theme }) => theme.spacing.spacing2};
+  padding: ${({ theme }) => theme.spacing.spacing2} ${({ theme }) => theme.spacing.spacing4};
+  cursor: pointer;
 `;
 
-export const FormRow = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-`;
-
-export const InputRow = styled.div`
+export const EmptyState = styled.div`
   display: flex;
   align-items: center;
-  width: 100%;
+  justify-content: center;
+  padding: ${({ theme }) => theme.spacing.spacing6};
+  border: 1px solid ${({ theme }) => theme.semantic.border.default};
+  border-radius: ${({ theme }) => theme.spacing.spacing2};
+  text-align: center;
 `;
 
-export const ErrorMessage = styled.span`
-  color: ${({ theme }) => theme.colors.red[700]};
-  ${({ theme }) => theme.typography.label2Regular};
-  margin-top: ${({ theme }) => theme.spacing.spacing1};
-  margin-left: calc(3.75rem + ${({ theme }) => theme.spacing.spacing3});
-  display: block;
-`;
-
-export const FormSpacer = styled.div`
-  height: ${({ theme }) => theme.spacing.spacing3};
+export const EmptyStateText = styled.p`
+  ${({ theme }) => theme.typography.body2Regular};
+  color: ${({ theme }) => theme.semantic.text.sub};
 `;
